Expose Ghidra's help() across the bridge.
Interactive mode now replaces help() with one that calls across the
bridge to access Ghidra's help. That only outputs via print (which can't
be replaced on Python2), so before calling it, it replaces sys.stdout
with one that has a write() method that captures the output when it's
called by print(), allowing it to be returned as a string
Interactive mode now replaces help() with one that calls across the
bridge to access Ghidra's help. That only outputs via print (which can't
be replaced on Python2), so before calling it, it replaces sys.stdout
with one that has a write() method that captures the output when it's
called by print(), allowing it to be returned as a string.